--- .prettierrc
@@ -1,9 +1,0 @@
-{
-  "semi": true,
-  "singleQuote": true,
-  "trailingComma": "all",
-  "printWidth": 80,
-  "tabWidth": 2,
-  "endOfLine": "crlf",
-  "plugins": ["prettier-plugin-tailwindcss"]
-}
