+++ src/components/ui/textarea.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../lib/utils";
+import { cn } from "../../shared/lib/utils";
 
 const Textarea = React.forwardRef<
   HTMLTextAreaElement,
