+++ src/types/scheduleTypes.ts
